Transfer microservice done

+++ MicroRabbit.Transfer.Domain/Models/TransferLog.cs
@@ -9,6 +9,6 @@ namespace MicroRabbit.Transfer.Domain.Models
         public int Id { get; set; }
         public int FromAccount { get; set; }
         public int ToAccount { get; set; }
-        public int TransferAmount { get; set; }
+        public decimal TransferAmount { get; set; }
     }
 }
